% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_your_bracket_fct_high_indvBracket.R
\name{fct_high_indvBracket}
\alias{fct_high_indvBracket}
\title{fct_high_indvBracket}
\usage{
fct_high_indvBracket(
  name,
  regions = c("West", "East", "South", "Midwest"),
  nodeColors = list(root = "white", unplayed = "black", won = "green", lost = "red"),
  scoring = 1,
  playerPicks,
  tournamentStandings,
  tournamentStructure,
  collapsed = FALSE
)
}
\arguments{
\item{name}{A character string of the name of the player}

\item{regions}{A character vector of the region names in the tourney}

\item{nodeColors}{A named list of colors. List names are: "root", "unplayed", "won", "lost"}

\item{scoring}{This is a place holder for a named list to be used to score players}

\item{playerPicks}{A named list containing the picks of every player in the pool}

\item{tournamentStandings}{A tibble of the current standings overlaid on top of the structure}

\item{tournamentStructure}{A tibble with the tournament structure and placeholders}

\item{collapsed}{Boolean of whether the graph should start expanded or collapsed.}
}
\value{
Here
}
\description{
Here
}
