% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_your_bracket_fct_low_prepBracket.R
\name{fct_low_prepBracket}
\alias{fct_low_prepBracket}
\title{fct_low_prepBracket}
\usage{
fct_low_prepBracket(input, region)
}
\arguments{
\item{input}{Any of the three input tibbles to transform}

\item{region}{A string of one of four of the regions in the tournament}
}
\value{
Named list
}
\description{
fct_low_prepBracket
}
