% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_your_bracket_fct_low_scorePicks.R
\name{fct_low_scorePicks}
\alias{fct_low_scorePicks}
\title{fct_low_scorePicks}
\usage{
fct_low_scorePicks(region, standings, picks, nodeColors, structure)
}
\arguments{
\item{region}{A string of one of four of the regions in the tournament}

\item{standings}{Test}

\item{picks}{Test}

\item{nodeColors}{Test}

\item{structure}{Test}
}
\value{
tibble
}
\description{
fct_low_scorePicks
}
